+++ mitra-json-signatures/src/verify.rs
@@ -195,7 +195,7 @@ mod tests {
             "type": "Test",
             "id": "https://example.org/objects/1",
             "proof": {
-                "type": "JcsEip191Signature2022",
+                "type": "MitraJcsEip191Signature2022",
                 "proofPurpose": "assertionMethod",
                 "verificationMethod": "did:pkh:eip155:1:0xb9c5714089478a327f09197987f16f9e5d936e8a",
                 "created": "2020-11-05T19:23:24Z",